alx_travel_app/requirements.txt
aiosqlite                     0.20.0     
amqp                          5.3.1      
asgiref                       3.8.1      
billiard                      4.2.1      
celery                        5.4.0      
certifi                       2024.12.14 
charset-normalizer            3.4.0
click                         8.1.8
click-didyoumean              0.3.1
click-plugins                 1.1.1
click-repl                    0.3.0
colorama                      0.4.6
Django                        5.0.8
django-cors-headers           4.6.0
django-environ                0.12.0
django-filter                 24.3
django-taggit                 6.0.0
djangorestframework           3.15.2
djangorestframework-simplejwt 5.3.1
drf-nested-routers            0.94.1
drf-yasg                      1.21.8
idna                          3.10
inflection                    0.5.1
kombu                         5.4.2
mysql-connector               2.2.9
mysql-connector-python        9.0.0
mysqlclient                   2.2.4
packaging                     24.2
parameterized                 0.9.0
pillow                        10.4.0
pip                           24.1.2
prompt_toolkit                3.0.50
pycodestyle                   2.12.1
PyJWT                         2.9.0
python-dateutil               2.9.0.post0
pytz                          2024.2
PyYAML                        6.0.2
requests                      2.32.3
six                           1.17.0
sqlparse                      0.5.1
typing_extensions             4.12.2
tzdata                        2024.1
uritemplate                   4.1.1
urllib3                       2.2.3
utils                         1.0.2
vine                          5.1.0
wcwidth                       0.2.13